split seller and buyer balances, buttons still in process
