+++ client/app/components/balance/Balance.less
@@ -101,7 +101,7 @@
       color: @dark-gray;
       border-bottom: 1px solid @border-gray;
       font-size: 20px;
-      padding: 24px 24px 24px 24px;
+      padding: 50px 24px 24px 24px;
       margin-bottom: 16px;
     }
   }
