+++ client/app/components/dashboard/MyAccount.tsx
@@ -5,11 +5,6 @@ import { storeConnectAcctToken,
          storeCustomerID,
          getConnectData } from '../../helpers/transactions';  
 
-import {  getBalances, 
-          toggleConfirm, 
-          toggleComplete, 
-          balanceDone } from '../../helpers/usersbalances'; 
-
 import { logout, getUserData } from '../../helpers/auth';
 import { HttpResponse, get, post, del } from '../../helpers/http';
 import './Dashboard.less';
@@ -31,7 +26,10 @@ const SideNav = () => {
       </div>
       <ul className="side-nav-list">
         <li className="nav-item">
-          <Link to="/dashboard">Current Balances</Link>
+          <Link to="/buying-balances">Buying Balances</Link>
+        </li>
+        <li className="nav-item">
+          <Link to="/selling-balances">Selling Balances</Link>
         </li>
         <li className="nav-item">
           <Link to="/history">History</Link>
