+++ client/app/components/dashboard/index.ts
@@ -1,6 +1,7 @@
-import Dashboard from './Dashboard';
+import Buying_Balances from './Buying_Balances';
+import Selling_Balances from './Selling_Balances';
 import History from './History';
 import BalanceSummary from './BalanceSummary';
 import MyAccount from './MyAccount';
 
-export { Dashboard, BalanceSummary, History, MyAccount };
+export { Buying_Balances, Selling_Balances, BalanceSummary, History, MyAccount };
