+++ client/app/components/nav/SideNav.tsx
@@ -12,8 +12,11 @@ const SideNav = () => {
 
       <ul className="side-nav-list">
         {/* TODO: make dynamic */}
-        <li className="nav-item active">
-          <a href="/dashboard">Current Balances</a>
+        <li className="nav-item">
+          <Link to="/buying-balances">Buying Balances</Link>
+        </li>
+        <li className="nav-item">
+          <Link to="/selling-balances">Selling Balances</Link>
         </li>
         <li className="nav-item">
           <a href="/history">History</a>
